Use modal for magic links
